initial paging implementation.
this implementation is based on the "cursor" approach; the cursor
parameter `c` records the lower or upper bounds for ordinals. the
presence of `c` implies that i) there was a non-empty page in the
reverse direction of the cursor and ii) there is a non-empty page
containing that ordinal. this information is enough for stateless
paging. note that the exact meaning of `c` is internal to the
server and should not be regarded permanent (I may try to make it
stable in the future however).